requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

# LangChain core pinned to 0.1.x to keep legacy imports working
langchain==0.1.16
langchain-community==0.0.32
langchain-google-genai==0.0.11

python-dotenv==1.0.1
faiss-cpu==1.8.0.post1

# Used by test_backend.py
streamlit==1.32.0